+++ src/components/form/checkbox-field.tsx
@@ -9,7 +9,7 @@ export type CheckboxFieldProps = {
 
 export function CheckboxField({ label, name, onChange, value }: CheckboxFieldProps) {
   return (
-    <div className="form-control hover:font-bold">
+    <div className="form-control">
       <label className="label cursor-pointer justify-start gap-1" htmlFor={name}>
         <input id={name} data-testid={kebabCase(name)} type="checkbox" className="checkbox checkbox-xs checkbox-primary" checked={value} onChange={event => onChange(event.target.checked)} />
         <span className="label-text text-sm tracking-tighter">{label}</span>