+++ src/utils/constants.ts
@@ -1,2 +1,1 @@
 export const maxPercentage = 100;
-export const nbDecimals = 2;