+++ src/utils/json.ts
@@ -18,16 +18,16 @@ export function jsonParse<Type>(jsonString: string): Type | undefined {
 const nbSpaces = 2;
 
 /**
- * Safely stringifies a JavaScript object into a JSON string. If stringification fails, it logs the error and returns an empty string.
+ * Safely stringifies a JavaScript object into a JSON string. If stringification fails, it logs the error and returns undefined.
  * @param data - The data to stringify.
- * @returns A JSON string representation of the data, or an empty string if stringification fails.
+ * @returns A JSON string representation of the data, or undefined if stringification fails.
  */
-export function jsonStringify(data: unknown): string {
+export function jsonStringify(data: unknown): string | undefined {
   try {
     return JSON.stringify(data, undefined, nbSpaces);
   } catch (error) {
     // oxlint-disable-next-line no-console
     console.error("Failed to stringify JSON :", error);
-    return "";
+    return undefined;
   }
 }